refactor: add loading state conditionally and work on feedbacks
