+++ src/App.tsx
@@ -1,5 +1,6 @@
+import DisplayTodos from "@/components/displayTodos/displayTodos";
+
 import "./App.css";
-import DisplayTodos from "./components/displayTodos";
 
 function App() {
   return <DisplayTodos />;
