+++ src/service/apiService.ts
@@ -1,0 +1,3 @@
+import axios from "axios";
+
+export const api = axios.create({ baseURL: "https://dummyjson.com" });
